refactor: move AI prompt types
