+++ backend/src/agents/news-analyst.ts
@@ -1,9 +1,15 @@
 import type { FastifyBaseLogger } from 'fastify';
 import { getNewsByToken } from '../repos/news.js';
 import { insertReviewRawLog } from '../repos/agent-review-raw-log.js';
-import { callAi, extractJson, type RebalancePrompt } from '../util/ai.js';
+import { callAi, extractJson } from '../util/ai.js';
 import { isStablecoin } from '../util/tokens.js';
-import { type AnalysisLog, type Analysis, analysisSchema, type RunParams } from './types.js';
+import {
+  type RebalancePrompt,
+  type AnalysisLog,
+  type Analysis,
+  analysisSchema,
+  type RunParams,
+} from './types.js';
 
 const CACHE_MS = 3 * 60 * 1000;
 const cache = new Map<
